ChatApp_by_AntiGravity/requirements.txt
fastapi
uvicorn
sqlalchemy
google-generativeai
python-dotenv
websockets
